pyproject.toml
[tool.poetry]
name = "doi2notion"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
crossrefapi = "^1.6.0"
python-dotenv = "^1.0.1"
arxiv = "^2.1.3"
tqdm = "^4.66.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.ruff]
extend-include = ["*.ipynb"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
